+++ src/Polysemy/Internal/Final.hs
@@ -5,43 +5,43 @@ module Polysemy.Internal.Final
     Final(..)
 
     -- * Actions
-  , withLoweringToFinal
+  , withTransToFinal
   , controlFinal
-  , withStrategicToFinal
+  , withLoweringToFinal
   , embedFinal
 
     -- * Combinators for Interpreting to the Final Monad
   , interpretFinal
 
-    -- * Strategy
-    -- | 'Strategic' is a domain-specific language very similar in
-    -- use to 'Polysemy.Interpretation.Handling', and is used to describe how
+    -- * Lowering
+    -- | 'Lowering' is a domain-specific language very similar in
+    -- use to 'Polysemy.HigherOrder', and is used to describe how
     -- higher-order effects are threaded down to the final monad.
     --
-    -- 'withStrategicToFinal' should be used when 'controlFinal' is
+    -- 'withLoweringToFinal' should be used when 'controlFinal' is
     -- not powerful enough for your purposes. Notable combinators include
-    -- 'runS', 'embed', 'liftWithS', and 'restoreS'.
-  , Strategic
+    -- 'runL', 'embed', 'liftWithL', and 'restoreL'.
+  , Lowering
 
-    -- ** Lifting the final monad to 'Strategic'
+    -- ** Lifting the final monad to 'Lowering'
   , embed
 
     -- ** Lowering computations to the final monad
-  , withProcessorS
-  , controlWithProcessorS
-  , processS
+  , withProcessorL
+  , controlWithProcessorL
+  , processL
 
-    -- ** Embedding computations into 'Strategic'
-  , runS
+    -- ** Embedding computations into 'Lowering'
+  , runL
 
-    -- ** Lowering 'Strategic' to the final monad
-  , controlS
-  , liftWithS
+    -- ** Lowering 'Lowering' to the final monad
+  , controlL
+  , liftWithL
 
     -- ** Manipulating effectful state
-  , restoreS
-  , runExposeS
-  , exposeS
+  , restoreL
+  , runExposeL
+  , exposeL
 
     -- * Interpretations
   , runM
@@ -55,7 +55,7 @@ import Control.Monad.Trans
 import Polysemy.Internal
 import Polysemy.Internal.Union
 import Polysemy.Internal.TH.Effect
-import Polysemy.Internal.Interpret (interpret)
+import Polysemy.Internal.HigherOrder (interpret)
 
 -----------------------------------------------------------------------------
 -- | An effect for embedding higher-order actions in the final target monad
@@ -69,7 +69,7 @@ import Polysemy.Internal.Interpret (interpret)
 -- If you only need the power of 'embed', then you should use 'Embed' instead.
 --
 -- /Note/: 'Final' actions are interpreted as actions of the final monad,
--- and the effectful state visible to 'controlFinal' \/ 'withStrategicToFinal'
+-- and the effectful state visible to 'controlFinal' \/ 'withLoweringToFinal'
 -- \/ 'interpretFinal' is that of /all interpreters run in order to produce the/
 -- /final monad/.
 --
@@ -89,78 +89,77 @@ import Polysemy.Internal.Interpret (interpret)
 --
 -- @since 1.2.0.0
 newtype Final m z a where
-  WithLoweringToFinal
+  WithTransToFinal
     :: (forall t. MonadTransWeave t => (forall x. z x -> t m x) -> t m a)
     -> Final m z a
 
 makeSem ''Final
 
--- | The 'Strategic' environment, which is a 'Polysemy.Sem' with a small
--- effect stack containing the @'Strategy' m t n@ effect, which provides the
+-- | The 'Lowering' environment, which is a 'Polysemy.Sem' with a small
+-- effect stack containing the @'Lower' m t n@ effect, which provides the
 -- various @-S@ combinators, as well as @'Embed' m@ and @'Final' m@ effects,
--- allowing you to embed actions of the final monad into 'Strategic'.
+-- allowing you to embed actions of the final monad into 'Lowering'.
 --
--- The type parameters of 'Strategic' represent the following:
+-- The type parameters of 'Lowering' represent the following:
 --
 -- * @m@: The final monad, the @m@ in @'Final' m@.
 -- * @t@: The type of final effectful state once /all/ interpreters -- past
 --   and future -- have been run, and @'Polysemy.Sem'@ has been reduced to the
 --   final monad.
 -- * @n@: The type of the \"source monad\": the monad that can be lowered t
---   @m@. When 'withStrategicToFinal' is used, @n@ is @Sem r@.
+--   @m@. When 'withLoweringToFinal' is used, @n@ is @Sem r@.
 --
 -- @since 2.0.0.0
-type Strategic m t n = Sem '[Strategy m t n, Embed m, Final m]
+type Lowering m t n = Sem '[Lower m t n, Embed m, Final m]
 
-data Strategy m t n z a where
-  LiftWithS :: forall m t n z a
-             . ((forall x. Strategic m t n x -> m (t x)) -> m a)
-            -> Strategy m t n z a
-  WithProcessorS :: forall m t n z a
+data Lower m t n z a where
+  LiftWithL :: forall m t n z a
+             . ((forall x. Lowering m t n x -> m (t x)) -> m a)
+            -> Lower m t n z a
+  WithProcessorL :: forall m t n z a
                   . ((forall x. n x -> m (t x)) -> m a)
-                 -> Strategy m t n z a
-  RestoreS  :: forall m t n z a. t a -> Strategy m t n z a
-  RunS      :: forall m t n z a. n a -> Strategy m t n z a
+                 -> Lower m t n z a
+  RestoreL  :: forall m t n z a. t a -> Lower m t n z a
+  RunL      :: forall m t n z a. n a -> Lower m t n z a
 
--- | Embed a computation of the source monad into 'Strategic'.
+-- | Embed a computation of the source monad into 'Lowering'.
 --
 -- @since 2.0.0.0
-runS :: forall m t n r a. n a -> Sem (Strategy m t n ': r) a
-runS = send . RunS @m @t
-{-# INLINE runS #-}
+runL :: forall m t n r a. n a -> Sem (Lower m t n ': r) a
+runL = send . RunL @m @t
+{-# INLINE runL #-}
 
 -- | Lift an computation of the final monad @m@ by giving it access to a
--- lowering function that can transform transform @'Strategic' m t n x@ to
--- @m (t x)@.
+-- lowering function that can transform @'Lowering' m t n x@ to @m (t x)@.
 --
 -- This is analogous to @liftBaseWith@ of @MonadBaseControl@.
 --
--- Note: the lowering function lowers @'Strategic' m t n@ by using the
--- effectful state as it is when 'liftWithS' is run.
+-- Note: the lowering function lowers @'Lowering' m t n@ by using the
+-- effectful state as it is when 'liftWithL' is run.
 --
 -- @since 2.0.0.0
-liftWithS :: forall m t n r a
-           . ((forall x. Strategic m t n x -> m (t x)) -> m a)
-          -> Sem (Strategy m t n ': r) a
-liftWithS main = send (LiftWithS main)
-{-# INLINE liftWithS #-}
+liftWithL :: forall m t n r a
+           . ((forall x. Lowering m t n x -> m (t x)) -> m a)
+          -> Sem (Lower m t n ': r) a
+liftWithL main = send (LiftWithL main)
+{-# INLINE liftWithL #-}
 
 -- | A particularly useful composition:
--- @'controlS' h = 'liftWithS' h >>= 'restoreS'@
+-- @'controlL' h = 'liftWithL' h >>= 'restoreL'@
 --
 -- This is analogous to @control@ of @MonadBaseControl@.
 --
--- Note: the lowering function lowers @'Strategic' m t n@ by using the
--- effectful state as it is when 'controlS' is run.
+-- Note: the lowering function lowers @'Lowering' m t n@ by using the
+-- effectful state as it is when 'controlL' is run.
 --
 -- @since 2.0.0.0
-controlS :: forall m t n r a
-          . ((forall x. Strategic m t n x -> m (t x)) -> m (t a))
-         -> Sem (Strategy m t n ': r) a
-controlS main = liftWithS main >>= restoreS
-{-# INLINE controlS #-}
+controlL :: forall m t n r a
+          . ((forall x. Lowering m t n x -> m (t x)) -> m (t a))
+         -> Sem (Lower m t n ': r) a
+controlL main = liftWithL main >>= restoreL
+{-# INLINE controlL #-}
 
--- | Embed a computation of the source monad into 'Strategic', and reify
+-- | Embed a computation of the source monad into 'Lowering', and reify
 --   the effectful state of all purely interpreted effects (effects not
 --   ultimately interpreted in terms of the final monad).
 --
@@ -176,12 +175,12 @@ controlS main = liftWithS main >>= restoreS
 --   never restoring the effectful state.
 --
 -- Once an effectful state has been reified, you may restore it using
--- 'restoreS'.
+-- 'restoreL'.
 --
 -- @since 2.0.0.0
-runExposeS :: forall m t n r a. n a -> Sem (Strategy m t n ': r) (t a)
-runExposeS z = withProcessorS $ \lower -> lower z
-{-# INLINE runExposeS #-}
+runExposeL :: forall m t n r a. n a -> Sem (Lower m t n ': r) (t a)
+runExposeL z = withProcessorL $ \lower -> lower z
+{-# INLINE runExposeL #-}
 
 -- | Restore a reified effectful state, bringing its changes into scope, and
 -- returning the result of the computation.
@@ -194,95 +193,95 @@ runExposeS z = withProcessorS $ \lower -> lower z
 -- For example, consider:
 --
 -- @
--- ta <- 'runExposeS' ma
--- tb <- 'runExposeS' mb
--- _  <- 'restoreS' ta
--- _  <- 'restoreS' tb
+-- ta <- 'runExposeL' ma
+-- tb <- 'runExposeL' mb
+-- _  <- 'restoreL' ta
+-- _  <- 'restoreL' tb
 -- @
 --
--- Unless @'restoreS' ta@ causes the handler to fail (because @ma@ failed due to
+-- Unless @'restoreL' ta@ causes the handler to fail (because @ma@ failed due to
 -- a local effect), the changes it brings into scope will be overridden by
--- @'restoreS' tb@.
+-- @'restoreL' tb@.
 --
 -- If you want to integrate the results of both actions, you need to restore the
--- state in between uses of 'runExposeS', so that @'runExposeS' mb@ works with
+-- state in between uses of 'runExposeL', so that @'runExposeL' mb@ works with
 -- the changes of @ta@ in scope.
 -- @
 -- ta <- 'runExposeH' ma
--- _  <- 'restoreS' ta
+-- _  <- 'restoreL' ta
 -- tb <- 'runExposeH' mb
 -- _  <- 'restoreH' tb
 -- @
 --
 -- @since 2.0.0.0
-restoreS :: forall m t n r a. t a -> Sem (Strategy m t n ': r) a
-restoreS = send . RestoreS @m @_ @n
-{-# INLINE restoreS #-}
+restoreL :: forall m t n r a. t a -> Sem (Lower m t n ': r) a
+restoreL = send . RestoreL @m @_ @n
+{-# INLINE restoreL #-}
 
 -- | Reify the effectful state of the purely interpreted effects used
 --   used within the argument.
 --
--- @'runExposeS' m = 'exposeS' ('runS' m)@
+-- @'runExposeL' m = 'exposeL' ('runL' m)@
 --
 -- @since 2.0.0.0
-exposeS :: forall m t n r a
-         . Strategic m t n a
-        -> Sem (Strategy m t n ': r) (t a)
-exposeS m = liftWithS $ \lower -> lower m
-{-# INLINE exposeS #-}
+exposeL :: forall m t n r a
+         . Lowering m t n a
+        -> Sem (Lower m t n ': r) (t a)
+exposeL m = liftWithL $ \lower -> lower m
+{-# INLINE exposeL #-}
 
 -- | Process a computation of the source monad by turning it into an computation
 -- of the final monad returning a reified effectful state. The processed
--- computation is returned, rather than being immediately run like with 'runS'.
+-- computation is returned, rather than being immediately run like with 'runL'.
 --
 -- /Note/: The processed action makes use of the effectful state as it is by
--- the time 'processS' is run, rather than what it is by the time the processed
+-- the time 'processL' is run, rather than what it is by the time the processed
 -- computation is run.
 --
 -- @since 2.0.0.0
-processS :: forall m t n r a
+processL :: forall m t n r a
           . Monad m
          => n a
-         -> Sem (Strategy m t n ': r) (m (t a))
-processS z = withProcessorS $ \lower -> return (lower z)
-{-# INLINE processS #-}
+         -> Sem (Lower m t n ': r) (m (t a))
+processL z = withProcessorL $ \lower -> return (lower z)
+{-# INLINE processL #-}
 
 -- | Lift an computation of the final monad by giving it access to a processor:
 -- a function that transforms a computation of the source monad by turning it
 -- into a computation of the final monad returning a reified effectful state.
 --
--- 'withProcessorS' is a useful specialization of 'liftWithS':
+-- 'withProcessorL' is a useful specialization of 'liftWithL':
 --
--- @'withProcessorS' main = 'liftWithS' (\n -> main (n . 'runS'))@
+-- @'withProcessorL' main = 'liftWithL' (\n -> main (n . 'runL'))@
 --
 -- /Note/: Processed actions makes use of the effectful state as it is by
--- the time 'withProcessorS' is run, rather than what it is by the time the
+-- the time 'withProcessorL' is run, rather than what it is by the time the
 -- processed action is run.
 --
 -- @since 2.0.0.0
-withProcessorS :: forall m t n r a
+withProcessorL :: forall m t n r a
                 . ((forall x. n x -> m (t x)) -> m a)
-               -> Sem (Strategy m t n ': r) a
-withProcessorS main = send (WithProcessorS main)
-{-# INLINE withProcessorS #-}
+               -> Sem (Lower m t n ': r) a
+withProcessorL main = send (WithProcessorL main)
+{-# INLINE withProcessorL #-}
 
 -- | A particularly useful composition:
--- @'controlWithProcessorS' h = 'withProcessorS' h >>= 'restoreS'@
+-- @'controlWithProcessorL' h = 'withProcessorL' h >>= 'restoreL'@
 --
--- 'withProcessorS' is a useful specialization of 'controlS':
+-- 'controlWithProcessorL' is a useful specialization of 'controlL':
 --
--- @'controlWithProcessorS' main = 'controlS' (\n -> main (n . 'runS'))@
+-- @'controlWithProcessorL' main = 'controlL' (\n -> main (n . 'runL'))@
 --
 -- /Note/: Processed actions makes use of the effectful state as it is by
--- the time 'controlWithProcessorS' is run, rather than what it is by the time
+-- the time 'controlWithProcessorL' is run, rather than what it is by the time
 -- the processed action is run.
 --
 -- @since 2.0.0.0
-controlWithProcessorS :: forall m t n r a
+controlWithProcessorL :: forall m t n r a
                        . ((forall x. n x -> m (t x)) -> m (t a))
-                      -> Sem (Strategy m t n ': r) a
-controlWithProcessorS main = withProcessorS main >>= restoreS
-{-# INLINE controlWithProcessorS #-}
+                      -> Sem (Lower m t n ': r) a
+controlWithProcessorL main = withProcessorL main >>= restoreL
+{-# INLINE controlWithProcessorL #-}
 
 -- | Lift an computation of the final monad by giving it access to a function
 -- that transforms any @'Sem' r@ computation into a compuation of the final
@@ -290,17 +289,17 @@ controlWithProcessorS main = withProcessorS main >>= restoreS
 -- monad must return a resulting effectful state.
 --
 -- The reified effectful state @t@ is 'Traversable', which gives you some
--- options -- see 'runExposeS'.
+-- options -- see 'runExposeL'.
 --
 -- 'controlFinal' provides no means of sequentially combining effectful states.
 -- If you need transform multiple @'Polysemy.Sem' r@ computations to the final
--- monad and sequentially execute them, then 'withStrategicToFinal' should be
+-- monad and sequentially execute them, then 'withLoweringToFinal' should be
 -- used instead.
 --
 -- You are discouraged from using 'controlFinal' in application code,
 -- as it ties your application code directly to the final monad.
 --
--- @'controlFinal' main = 'withStrategicToFinal' ('controlWithProcessorS' main)@
+-- @'controlFinal' main = 'withLoweringToFinal' ('controlWithProcessorL' main)@
 --
 -- @since 2.0.0.0
 controlFinal :: forall m r a
@@ -310,27 +309,27 @@ controlFinal :: forall m r a
                 => (forall x. Sem r x -> m (t x)) -> m (t a)
                 )
              -> Sem r a
-controlFinal main = withLoweringToFinal $ \n ->
+controlFinal main = withTransToFinal $ \n ->
   controlT $ \lower -> main (lower . n)
 {-# INLINE controlFinal #-}
 
-runStrategy :: forall m n t a
+runLowering :: forall m n t a
              . (Monad m, MonadTransWeave t)
-            => Strategic m (StT t) n a
+            => Lowering m (StT t) n a
             -> (forall x. n x -> t m x) -> t m a
-runStrategy main nat =
+runLowering main nat =
   let
-    go :: forall x. Strategic m (StT t) n x -> t m x
+    go :: forall x. Lowering m (StT t) n x -> t m x
     go = usingSem $ \(Union pr (Weaving eff mkT lwr ex)) -> do
       let run_it = (ex . (<$ mkInitState lwr))
       case pr of
         Here -> run_it <$> case eff of
-          RestoreS t -> restoreT (return t)
-          RunS m -> nat m
-          LiftWithS main' -> liftWith $ \lower -> main' (lower . go)
-          WithProcessorS main' -> liftWith $ \lower -> main' (lower . nat)
+          RestoreL t -> restoreT (return t)
+          RunL m -> nat m
+          LiftWithL main' -> liftWith $ \lower -> main' (lower . go)
+          WithProcessorL main' -> liftWith $ \lower -> main' (lower . nat)
         There Here | Embed m <- eff -> run_it <$> lift m
-        There (There Here) | WithLoweringToFinal main' <- eff ->
+        There (There Here) | WithTransToFinal main' <- eff ->
           fmap ex $ lwr $ getComposeT $ main' (ComposeT . mkT go)
         There (There (There pr_)) -> case pr_ of {}
   in
@@ -339,19 +338,19 @@ runStrategy main nat =
 -----------------------------------------------------------------------------
 -- | Allows for embedding higher-order actions of the final monad
 -- by providing the means of explicitly threading effects through @'Sem' r@
--- to the final monad. This is done through the use of the 'Strategy'
+-- to the final monad. This is done through the use of the 'Lower'
 -- environment, which provides a variety of combinators, most notably
--- 'controlS'.
+-- 'controlL'.
 --
--- You are discouraged from using 'withStrategicToFinal' in application code,
+-- You are discouraged from using 'withLoweringToFinal' in application code,
 -- as it ties your application code directly to the final monad.
 --
 -- @since 2.0.0.0
-withStrategicToFinal :: (Monad m, Member (Final m) r)
-                     => (forall t. Traversable t => Strategic m t (Sem r) a)
+withLoweringToFinal :: (Monad m, Member (Final m) r)
+                     => (forall t. Traversable t => Lowering m t (Sem r) a)
                      -> Sem r a
-withStrategicToFinal main = withLoweringToFinal (runStrategy main)
-{-# INLINE withStrategicToFinal #-}
+withLoweringToFinal main = withTransToFinal (runLowering main)
+{-# INLINE withLoweringToFinal #-}
 
 ------------------------------------------------------------------------------
 -- | Lower a 'Sem' containing only a single lifted 'Monad' into that
@@ -360,19 +359,19 @@ runM :: Monad m => Sem '[Embed m, Final m] a -> m a
 runM = usingSem $ \u -> case decomp u of
   Right (Weaving (Embed m) _ lwr ex) -> fmap (ex . (<$ mkInitState lwr)) m
   Left g -> case extract g of
-    Weaving (WithLoweringToFinal main) mkT lwr ex ->
+    Weaving (WithTransToFinal main) mkT lwr ex ->
       fmap ex $ lwr $ main $ mkT runM
 {-# INLINE runM #-}
 
 
 -----------------------------------------------------------------------------
--- | 'withStrategicToFinal' admits an implementation of 'embed'.
+-- | 'withLoweringToFinal' admits an implementation of 'embed'.
 --
 -- Just like 'embed', you are discouraged from using this in application code.
 --
 -- @since 1.2.0.0
 embedFinal :: (Member (Final m) r, Monad m) => m a -> Sem r a
-embedFinal m = withLoweringToFinal $ \_ -> lift m
+embedFinal m = withTransToFinal $ \_ -> lift m
 {-# INLINE embedFinal #-}
 
 ------------------------------------------------------------------------------
@@ -380,16 +379,15 @@ embedFinal m = withLoweringToFinal $ \_ -> lift m
 -- interpret higher-order effects in terms of the final monad.
 --
 -- 'interpretFinal' requires less boilerplate than using 'interpretH'
--- together with 'withStrategicToFinal' \/ 'withWeavingToFinal',
--- but is also less powerful.
+-- together with 'withLoweringToFinal', but is also less powerful.
 -- 'interpretFinal' does not provide any means of executing actions
 -- of @'Sem' r@ as you interpret each action, and the provided interpreter
 -- is automatically recursively used to process higher-order occurences of
 -- @'Sem' (e ': r)@ to @'Sem' r@.
 --
 -- If you need greater control of how the effect is interpreted,
--- use 'interpretH' together with 'withStrategicToFinal' \/
--- 'withWeavingToFinal' instead.
+-- use 'interpretH' together with 'withLoweringToFinal' \/
+-- 'controlFinal' instead.
 --
 -- /Note/: Effects that aren't interpreted in terms of the final
 -- monad will have local state semantics in regards to effects
@@ -399,8 +397,8 @@ embedFinal m = withLoweringToFinal $ \_ -> lift m
 interpretFinal
     :: forall m e r a
      . (Member (Final m) r, Monad m)
-    => (forall t z x. Traversable t => e z x -> Strategic m t z x)
-       -- ^ A handler written using the 'Strategic' environment, where
+    => (forall t z x. Traversable t => e z x -> Lowering m t z x)
+       -- ^ A handler written using the 'Lowering' environment, where
        -- the source monad @z@ is the monad of the higher-order chunks in @e@.
     -> Sem (e ': r) a
     -> Sem r a
@@ -411,7 +409,7 @@ interpretFinal h =
       Right (Weaving e mkT lwr ex) ->
         injWeaving $
           Weaving
-            (WithLoweringToFinal (runStrategy (h e)))
+            (WithTransToFinal (runLowering (h e)))
             (\n -> mkT (n . go))
             lwr
             ex
@@ -436,11 +434,10 @@ finalToFinal to from =
   let
     go :: Sem (Final m1 ': r) x -> Sem r x
     go = hoistSem $ \u -> case decomp u of
-      Right (Weaving (WithLoweringToFinal main) mkT lwr ex) ->
+      Right (Weaving (WithTransToFinal main) mkT lwr ex) ->
         injWeaving $
           Weaving
-            (WithLoweringToFinal $ \n -> hoistT to $ main (hoistT from . n)
-            )
+            (WithTransToFinal $ \n -> hoistT to $ main (hoistT from . n))
             (\n -> mkT (n . go))
             lwr
             ex